+++ src/model/RecipeModel.js
@@ -2,7 +2,7 @@ const Pool = require('../config/db')
 
 const getRecipeAll = async ({ searchRecipe, searchBy, sortBy, sort, offset, limit }) => {
     return new Promise((resolve, reject) =>
-        Pool.query(`SELECT recipe.id, recipe.title, recipe.ingredients, recipe.photo, category.category_name AS category FROM recipe JOIN category ON recipe.category_id = category.category_id  WHERE ${searchBy} ILIKE '%${searchRecipe}%' ORDER BY ${sortBy} ${sort} OFFSET ${offset} LIMIT ${limit}`, (err, result) => {
+        Pool.query(`SELECT recipe.id, recipe.title, recipe.ingredients, recipe.photo, category.category_name AS category, recipe.user_id FROM recipe JOIN category ON recipe.category_id = category.category_id  WHERE ${searchBy} ILIKE '%${searchRecipe}%' ORDER BY ${sortBy} ${sort} OFFSET ${offset} LIMIT ${limit}`, (err, result) => {
             if (!err) {
                 resolve(result)
             } else {
@@ -15,6 +15,30 @@ const getRecipeAll = async ({ searchRecipe, searchBy, sortBy, sort, offset, limi
 const getRecipeAllCount = async () => {
     return new Promise((resolve, reject) =>
         Pool.query(`SELECT * FROM recipe `, (err, result) => {
+            if (!err) {
+                resolve(result)
+            } else {
+                reject(err)
+            }
+        })
+    )
+}
+
+const getRecipeAllByUserId = async ({ userid, offset, limit }) => {
+    return new Promise((resolve, reject) =>
+        Pool.query(`SELECT recipe.id, recipe.title, recipe.ingredients, recipe.photo, category.category_name AS category, recipe.user_id FROM recipe JOIN category ON recipe.category_id = category.category_id  WHERE recipe.user_id='${userid}' OFFSET ${offset} LIMIT ${limit}`, (err, result) => {
+            if (!err) {
+                resolve(result)
+            } else {
+                reject(err)
+            }
+        })
+    )
+}
+
+const getRecipeAllByUserIdCount = async ({ userid }) => {
+    return new Promise((resolve, reject) =>
+        Pool.query(`SELECT * FROM recipe where user_id = '${userid}'`, (err, result) => {
             if (!err) {
                 resolve(result)
             } else {
@@ -76,4 +100,4 @@ const deleteById = async (id) => {
 }
 
 
-module.exports = { getRecipeAll, getRecipeAllCount, getRecipeById, postRecipe, putRecipe, deleteById }
+module.exports = { getRecipeAll, getRecipeAllCount, getRecipeAllByUserId, getRecipeAllByUserIdCount, getRecipeById, postRecipe, putRecipe, deleteById }
